LK-Call-Initiator/pyproject.toml
[project]
name = "lk-call-initiator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "livekit-agents>=1.3.10",
    "python-dotenv>=1.2.1",
]

[tool.uv]
environments = [
    "sys_platform == 'linux' and platform_machine == 'x86_64'",
    "sys_platform == 'darwin' and platform_machine == 'arm64'", # Add if you use Mac M1/M2/M3
    "sys_platform == 'win32' and platform_machine == 'AMD64'", # Add if you use Windows
]
